Simplify docker build steps
Remove unnecessary manual build instructions that's now done automatically in the build step in the `Dockerfile`
Added `-d` flag to the docker command to make it run in the background by default. 
Added notes on how to inspect the docker container properly and updated the note about the `-d` tag.